// src/types.d.ts
export interface DtsGenerationOption {cwd?: string
  tsconfigPath?: string
  root?: string
  outdir?: string
  keepComments?: boolean}
export interface DtsGenerationOption {cwd?: string
  tsconfigPath?: string
  root?: string
  outdir?: string
  keepComments?: boolean}
export interface DtsGenerationOption {cwd?: string
  tsconfigPath?: string
  root?: string
  outdir?: string
  keepComments?: boolean}